// src/app/WebClient/Clients/interface/catalog.interface.ts
// Generated by https://quicktype.io

export interface Product {
  hasItems: boolean;
  items: ProductItem[];
  total: number;
  page: number;
  pages: number;
}

export interface ProductItem {
  productId: number;
  name: string;
  description: string;
  price: number;
  stock: number;
}



export interface ProductInStock {
  hasItems: boolean;
  items: ProductInStockItem[];
  total: number;
  page: number;
  pages: number;
}

export interface ProductInStockItem {
  productInStockId: number;
  productId: number;
  stock: number;
}